chore(be): address aterga's review comments
* MAX_VERIFIED_EMAILS_PER_ANCHOR is now `u8` instead of `usize`, so
  the LimitReached { limit: u8 } variant no longer needs an `as u8`
  cast. Comparison sites use `usize::from(...)` (infallible widen).
* Rename `verified_emails::prepare_add` →
  `prepare_add_verified_email` and `email_inbound::smtp::bind_verified_email`
  → `append_or_refresh_verified_email`; rename `mut a` → `mut anchor`
  inside the latter.
* `add_openid_credential[_skip_checks]` now take an explicit `now_ns`
  parameter (callers pass `ic_cdk::api::time()`), giving the verified-
  emails cap-on-OIDC-link path a testable seam.
* New unit test covering the requested scenario: linking an OIDC
  credential when the anchor already has the full
  MAX_VERIFIED_EMAILS_PER_ANCHOR verified emails — the link itself
  succeeds, the silent mirror is skipped, and the verified-emails
  bucket doesn't grow.

+++ src/internet_identity/src/email_inbound/mod.rs
@@ -101,7 +101,7 @@ pub const VERIFIED_EMAIL_NONCE_PREFIX: &str = "II-Verify-";
 /// users will have at most one or two; 5 leaves clear headroom for
 /// work/personal/alias combinations without inviting credential
 /// stuffing.
-pub const MAX_VERIFIED_EMAILS_PER_ANCHOR: usize = 5;
+pub const MAX_VERIFIED_EMAILS_PER_ANCHOR: u8 = 5;
 
 /// Byte length of the random suffix appended after [`NONCE_PREFIX`].
 /// 8 bytes → 16 hex chars → ~64 bits of entropy. The challenge map
